src/backend-users/requirements.txt
attrs==23.2.0
blinker==1.7.0
cattrs==23.2.3
click==8.1.7
exceptiongroup==1.2.0
Flask==3.0.3
flask-swagger-ui==4.11.1
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.3
lsprotocol==2023.0.1
MarkupSafe==2.1.5
msgpack==1.0.7
numpy==1.26.2
psycopg2-binary==2.9.9
pygls==1.3.0
pynvim==0.5.0
python-dotenv==1.0.1
SQLAlchemy==2.0.29
typing_extensions==4.9.0
Werkzeug==3.0.2
